+++ pool/app/session/repo/repo.ml
@@ -559,10 +559,10 @@ module Sql = struct
   let find_for_calendar_by_location_request =
     let open Caqti_request.Infix in
     {sql|
-        pool_sessions.location_uuid = UNHEX(REPLACE($1, '-', ''))
-        AND pool_sessions.canceled_at IS NULL
+        pool_sessions.canceled_at IS NULL
         AND pool_sessions.start > $2
         AND pool_sessions.start < $3
+        AND pool_sessions.location_uuid = UNHEX(REPLACE($1, '-', ''))
       |sql}
     |> select_for_calendar ~order_by:"pool_sessions.start"
     |> Caqti_type.(tup3 string ptime ptime ->* RepoEntity.Calendar.t)
@@ -575,35 +575,29 @@ module Sql = struct
       (Pool_location.Id.value location_id, start_time, end_time)
   ;;
 
-  let validate_session_sql actor =
-    let open Guard.Persistence in
-    let open Dynparam in
-    ( {sql| guardianValidateSessionUuid(guardianEncodeUuid(?), ?, pool_sessions.uuid) |sql}
-    , empty
-      |> add Uuid.Actor.t (actor |> Guard.Actor.id)
-      |> add Action.t Guard.Action.Read )
-  ;;
-
   let find_for_calendar_by_user_request =
     select_for_calendar ~order_by:"pool_sessions.start"
   ;;
 
   let find_for_calendar_by_user actor pool ~start_time ~end_time =
     let open Caqti_request.Infix in
-    let sql, dyn = validate_session_sql actor in
     let sql =
-      Format.asprintf
-        "%s AND %s AND %s AND %s"
-        sql
-        "pool_sessions.start > ?"
-        "pool_sessions.start < ?"
-        "pool_sessions.canceled_at IS NULL"
+      [ "pool_sessions.start > ?"
+      ; "pool_sessions.start < ?"
+      ; "pool_sessions.canceled_at IS NULL"
+      ; {sql|guardianValidateExperimentUuid(guardianEncodeUuid(?), ?, pool_experiments.uuid)|sql}
+      ]
+      |> CCString.concat " AND "
     in
     let dyn =
+      let open Guard.Persistence in
+      let open Dynparam in
       CCList.fold_left
         (fun dyn p -> dyn |> Dynparam.add Caqti_type.ptime p)
-        dyn
+        empty
         [ start_time; end_time ]
+      |> add Uuid.Actor.t (actor |> Guard.Actor.id)
+      |> add Action.t Guard.Action.Read
     in
     let (Dynparam.Pack (pt, pv)) = dyn in
     let request =