+++ pool/test/session_test.ml
@@ -42,13 +42,13 @@ module Data = struct
     let start1 = Raw.start1 |> Ptime.to_rfc3339 ~frac_s:12
     let start2 = Raw.start2 |> Ptime.to_rfc3339 ~frac_s:12
     let start3 = Raw.start3 |> Ptime.to_rfc3339 ~frac_s:12
-    let duration = Raw.duration |> Pool_common.Utils.Time.timespan_to_hours
+    let duration = Raw.duration |> Pool_common.Utils.Time.timespan_to_minutes
     let description = Raw.description
     let limitations = Raw.limitations
     let max_participants = Raw.max_participants |> string_of_int
     let min_participants = Raw.min_participants |> string_of_int
     let overbook = Raw.overbook |> string_of_int
-    let lead_time = Raw.lead_time |> Pool_common.Utils.Time.timespan_to_hours
+    let lead_time = Raw.lead_time |> Pool_common.Utils.Time.timespan_to_minutes
     let sent_at = Raw.sent_at |> Ptime.to_rfc3339 ~frac_s:12
     let assignment_count = Raw.assignment_count |> string_of_int
   end