+++ resources/calendar.js
@@ -4,6 +4,7 @@ import timeGridPlugin from '@fullcalendar/timegrid';
 import listPlugin from '@fullcalendar/list';
 import tippy from 'tippy.js';
 import { notifyUser } from "./admin/utils"
+import { generateColor } from "./utils/color"
 
 const notificationId = "calendar-notification";
 
@@ -18,8 +19,13 @@ const toLocalTime = date => date.toLocaleTimeString('en',
     })
 
 const normalizeSession = (session) => {
+    const color = generateColor(session.title)
     session.end = session.end_
     delete session.end_
+    session.textColor = "#363636"
+    session.backgroundColor = `${color}${Math.floor(0.2 * 255).toString(16)}`
+    session.borderColor = color
+    session.display = "block"
     return session
 }